requirements.txt
matplotlib==3.3.4
pillow==8.1.0
pandas==1.2.2
seaborn==0.11.1

